C2PA & GEMINI/ai-forensics-tool/requirements.txt
Flask==3.0.0
fpdf2==2.7.9
Pillow==10.3.0
numpy==1.26.4
